(PC-8914) Rename YoungerBagde to ProfileBadge
